Consolidate excavate_ite tests into the main module
Move the test cases from the separate excavate_ite/tests/{bitvec,bool}_test.rs
files into an inline `#[cfg(test)] mod tests` in excavate_ite.rs, matching the
convention used by the other algorithm modules (dfs, canonicalize, collect_vars,
etc.). Removes the now-empty tests submodule directory.
