--- crates/clarirs-core/src/algorithms/excavate_ite/tests.rs
@@ -1,2 +1,0 @@
-mod bitvec_test;
-mod bool_test;
